Add common telemetry
