+++ src/common/utils.ts
@@ -3,102 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { EventEmitter, Event, Disposable } from 'vscode';
-
-export function filterEvent<T>(event: Event<T>, filter: (e: T) => boolean): Event<T> {
-	return (listener, thisArgs = null, disposables?) => event(e => filter(e) && listener.call(thisArgs, e), null, disposables);
-}
-
-/**
- * Given an event, returns another event which only fires once.
- *
- * @param event The event source for the new event.
- */
-export function onceEvent<T>(event: Event<T>): Event<T> {
-	return (listener, thisArgs = null, disposables?) => {
-		// we need this, in case the event fires during the listener call
-		let didFire = false;
-		let result: Disposable | undefined = undefined;
-		result = event(e => {
-			if (didFire) {
-				return;
-			} else if (result) {
-				result.dispose();
-			} else {
-				didFire = true;
-			}
-
-			return listener.call(thisArgs, e);
-		}, null, disposables);
-
-		if (didFire) {
-			result.dispose();
-		}
-
-		return result;
-	};
-}
-
-export interface PromiseAdapter<T, U> {
-	(
-		value: T,
-		resolve:
-			(value: U | PromiseLike<U>) => void,
-		reject:
-			(reason: any) => void
-	): any;
-}
-
-const passthrough = (value: any, resolve: (value?: any) => void) => resolve(value);
-
-/**
- * Return a promise that resolves with the next emitted event, or with some future
- * event as decided by an adapter.
- *
- * If specified, the adapter is a function that will be called with
- * `(event, resolve, reject)`. It will be called once per event until it resolves or
- * rejects.
- *
- * The default adapter is the passthrough function `(value, resolve) => resolve(value)`.
- *
- * @param event the event
- * @param adapter controls resolution of the returned promise
- * @returns a promise that resolves or rejects as specified by the adapter
- */
-export function promiseFromEvent<T, U>(
-	event: Event<T>,
-	adapter: PromiseAdapter<T, U> = passthrough): { promise: Promise<U>; cancel: EventEmitter<void> } {
-	let subscription: Disposable;
-	let cancel = new EventEmitter<void>();
-	return {
-		promise: new Promise<U>((resolve, reject) => {
-			cancel.event(_ => reject('Cancelled'));
-			subscription = event((value: T) => {
-				try {
-					Promise.resolve(adapter(value, resolve, reject))
-						.catch(reject);
-				} catch (error) {
-					reject(error);
-				}
-			});
-		}).then(
-			(result: U) => {
-				subscription.dispose();
-				return result;
-			},
-			error => {
-				subscription.dispose();
-				throw error;
-			}
-		),
-		cancel
-	};
-}
-
-export function eventToPromise<T>(event: Event<T>): Promise<T> {
-	return new Promise(resolve => onceEvent(event)(resolve));
-}
-
 export function arrayEquals<T>(one: ReadonlyArray<T> | undefined, other: ReadonlyArray<T> | undefined, itemEquals: (a: T, b: T) => boolean = (a, b) => a === b): boolean {
 	if (one === other) {
 		return true;
